pyproject.toml
[tool.poetry]
name = "jas"
version = "0.1.2"
description = "Jabber Auto Update Server API"
authors = ["Ryan Mullins"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.85.0"
uvicorn = "^0.18.3"
pyyaml = "^6.0"
python-multipart = "^0.0.5"
lxml = "^4.9.1"
pydantic = {extras = ["dotenv"], version = "^1.10.2"}
gunicorn = "^20.1.0"

[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
coverage = {extras = ["toml"], version = "^6.5.0"}
isort = {extras = ["pyproject"], version = "^5.10.1"}
pytest = "^7.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
